@@ -1,79 +1,50 @@
 {
-  "_args": [
-    [
-      "normalize-url@4.5.0",
-      "C:\\GitHub\\send-to-allure-server-action"
-    ]
-  ],
-  "_from": "normalize-url@4.5.0",
-  "_id": "normalize-url@4.5.0",
-  "_inBundle": false,
-  "_integrity": "sha512-2s47yzUxdexf1OhyRi4Em83iQk0aPvwTddtFz4hnSSw9dCEsLEGf6SwIO8ss/19S9iBb5sJaOuTvTGDeZI00BQ==",
-  "_location": "/normalize-url",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "normalize-url@4.5.0",
-    "name": "normalize-url",
-    "escapedName": "normalize-url",
-    "rawSpec": "4.5.0",
-    "saveSpec": null,
-    "fetchSpec": "4.5.0"
-  },
-  "_requiredBy": [
-    "/cacheable-request"
-  ],
-  "_resolved": "https://registry.npmjs.org/normalize-url/-/normalize-url-4.5.0.tgz",
-  "_spec": "4.5.0",
-  "_where": "C:\\GitHub\\send-to-allure-server-action",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "bugs": {
-    "url": "https://github.com/sindresorhus/normalize-url/issues"
-  },
-  "description": "Normalize a URL",
-  "devDependencies": {
-    "ava": "^2.4.0",
-    "coveralls": "^3.0.6",
-    "nyc": "^14.1.1",
-    "tsd": "^0.8.0",
-    "xo": "^0.24.0"
-  },
-  "engines": {
-    "node": ">=8"
-  },
-  "files": [
-    "index.js",
-    "index.d.ts"
-  ],
-  "homepage": "https://github.com/sindresorhus/normalize-url#readme",
-  "keywords": [
-    "normalize",
-    "url",
-    "uri",
-    "address",
-    "string",
-    "normalization",
-    "normalisation",
-    "query",
-    "querystring",
-    "simplify",
-    "strip",
-    "trim",
-    "canonical"
-  ],
-  "license": "MIT",
-  "name": "normalize-url",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sindresorhus/normalize-url.git"
-  },
-  "scripts": {
-    "test": "xo && nyc ava && tsd"
-  },
-  "version": "4.5.0"
+	"name": "normalize-url",
+	"version": "6.1.0",
+	"description": "Normalize a URL",
+	"license": "MIT",
+	"repository": "sindresorhus/normalize-url",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"engines": {
+		"node": ">=10"
+	},
+	"scripts": {
+		"test": "xo && nyc ava && tsd"
+	},
+	"files": [
+		"index.js",
+		"index.d.ts"
+	],
+	"keywords": [
+		"normalize",
+		"url",
+		"uri",
+		"address",
+		"string",
+		"normalization",
+		"normalisation",
+		"query",
+		"querystring",
+		"simplify",
+		"strip",
+		"trim",
+		"canonical"
+	],
+	"devDependencies": {
+		"ava": "^2.4.0",
+		"nyc": "^15.0.0",
+		"tsd": "^0.11.0",
+		"xo": "^0.25.3"
+	},
+	"nyc": {
+		"reporter": [
+			"text",
+			"lcov"
+		]
+	}
 }
